-Adds GET /api/imdb/search/:movie
-Uses Cheerio to scrape IMDB
-Responds with body in JSON format
