--- api/routes.js
@@ -1,5 +1,0 @@
-const express = require('express')
-const router = express.Router()
-
-
-module.exports = router;
