+++ package.json
@@ -2,9 +2,10 @@
   "name": "moviesearchapi",
   "version": "1.0.0",
   "description": "Using Express, create a web server that exposes a Movie Search HTTP API which fetches results by scraping imdb.com.",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "nodemon"
   },
   "repository": {
     "type": "git",
@@ -21,13 +22,14 @@
     "babel": "^6.23.0",
     "babel-core": "^6.26.0",
     "babel-preset-env": "^1.1.4",
+    "body-parser": "^1.18.2",
     "chai": "^4.1.1",
     "cheerio": "^1.0.0-rc.2",
     "cheerio-tableparser": "^1.0.1",
+    "express": "^4.15.5",
     "mocha": "^3.5.0",
     "nock": "^9.0.14",
     "request": "^2.81.0",
     "request-promise": "^4.2.1"
   }
 }
-
